+++ src/index.ts
@@ -2,5 +2,7 @@
 // Exports
 //
 
+export * from "./data/local-ip-ranges.js";
+
 export * from "./functions/escape-sql-like-string.js";
 export * from "./functions/split-string-into-chunks.js";