+++ src/index.ts
@@ -3,6 +3,7 @@
 //
 
 export * from "./data/local-ip-ranges.js";
+export * from "./data/random-string-characters.js";
 
 export * from "./functions/escape-sql-like-string.js";
 export * from "./functions/is-local-ip-address.js";