+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 ## 1.1.0
-Added `splitStringIntoChunks` function.
+
+* Added `splitStringIntoChunks` function.
+* Added `localIpRanges` array.
 
 ## 1.0.0
 Initial release.