+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 * Added `splitStringIntoChunks` function.
 * Added `localIpRanges` array.
+* Added `isLocalIpAddress` function.
 
 ## 1.0.0
 Initial release.